♻️ Refactor: ErrorBoundary에서 404에러 분리
# 커밋 메시지 형식
# <이모지><타입>: <제목> (#이슈 번호)
#
# ⛔ 제목은 100자 이내, 끝에 마침표(.) 사용 금지
# ⛔ 이모지와 타입 사이에 공백 없이 붙여 씁니다 (예: ✨Feat)
# ⛔ 제목과 콜론(:) 사이에 공백 없이 붙이고, 콜론 뒤에는 공백 1칸 필요

# ✅ 제목은 커밋의 "무엇"을 나타냅니다
# ✅ 본문은 "왜", "어떻게"를 요약하며, 선택사항입니다
# ✅ 본문은 여러 줄의 메시지를 작성할 땐 "-"로 구분

# 예시:
# ✨Feat: 마이페이지 수정 기능 추가 (#15)

# 커밋 타입 가이드:
# ✨ Feat     : 새로운 기능 추가
# 🐛 Fix      : 버그 수정
# 📦 Chore    : 설정/빌드 수정, 기타 잡일
# 💄 Style    : 포맷팅, 세미콜론 등 비즈니스 로직 없는 변경
# 📝 Docs     : 문서 작성 및 수정
# ♻️ Refactor : 리팩토링 (동작 변경 없이 코드 개선)
