--- src/pages/dev/IdolRegisterPage.jsx
@@ -1,1 +1,0 @@
-// 상의 필요. 폼에서 입력한 값들을 서버에 전송하는 부분을 구현.. 버튼 클릭 시 서버에 전송
