+++ src/pages/list/ListPage.jsx
@@ -12,7 +12,7 @@ export const wrapper = css`
 `;
 
 export const carousel = css`
-  margin: 12rem 0
+  margin: 12rem 0;
 `;
 
 function ListPage() {
