+++ src/routes/router.jsx
@@ -1,7 +1,8 @@
 import App from '@/App.jsx';
-import ErrorBoundary from '@/components/ErrorBoundary.jsx';
-import PendingUI from '@/components/PendingUI.jsx';
+import ErrorBoundary from '@/components/feedback/ErrorBoundary.jsx';
+import PendingUI from '@/components/feedback/PendingUI.jsx';
 import { ENDPOINTS } from '@/constants/api';
+import NotFoundPage from '@/pages/error/NotFoundPage';
 import fetchData from '@/utils/fetchData';
 import { lazy } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
@@ -61,6 +62,10 @@ const router = createBrowserRouter([
         errorElement: <ErrorBoundary />,
       },
     ],
+  },
+  {
+    path: '*',
+    element: <NotFoundPage />,
   },
 ]);
